dataloaders/pyproject.toml
[project]
name = "Dataloaders"
version = "0.1.0"
description = "Python package and scripts with useful dataloaders"
readme = "README.md"
requires-python = ">=3.10,<=3.13.3"
dependencies = [
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "MNIST_dataloaders",
]
